requirements.txt
torch==1.12.0
torchvision==0.13.0
Pillow==9.3.0
opencv-python == 4.6.0.66
numpy==1.23.4
python-multipart==0.0.5